Moved FindCourseSelectionController
