--- app/controllers/candidate_interface/continuous_applications/course_choices/find_course_selection_controller.rb
@@ -1,16 +1,0 @@
-module CandidateInterface
-  module ContinuousApplications
-    module CourseChoices
-      class FindCourseSelectionController < CandidateInterface::CourseChoices::BaseController
-        include CandidateInterface::CourseChoices::Concerns::DuplicateCourseRedirect
-        include CandidateInterface::CourseChoices::Concerns::FullCourseRedirect
-
-      private
-
-        def current_step
-          :find_course_selection
-        end
-      end
-    end
-  end
-end
