+++ app/controllers/candidate_interface/course_choices/find_course_selection_controller.rb
@@ -1,0 +1,14 @@
+module CandidateInterface
+  module CourseChoices
+    class FindCourseSelectionController < CandidateInterface::CourseChoices::BaseController
+      include CandidateInterface::CourseChoices::Concerns::DuplicateCourseRedirect
+      include CandidateInterface::CourseChoices::Concerns::FullCourseRedirect
+
+    private
+
+      def current_step
+        :find_course_selection
+      end
+    end
+  end
+end
