+++ app/views/candidate_interface/course_choices/find_course_selection/new.html.erb
